pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "celery>=5.6.3",
    "fastapi>=0.136.1",
    "inngest>=0.5.18",
    "llama-index>=0.14.21",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-openai>=0.7.5",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
